third_party/blink/web_tests/fast/forms/select/select-selectedOptions-expected.txt

one
two
1) Initial there is no selected options.
PASS optionsLength is 2
PASS selectedOptions.length is 0
2) Select an option should update the selected options collection.
PASS optionsLength is 2
PASS selectedOptions.length is 1
PASS selectedOptions[0].text is 'one'
3) Select two options should update the selected options collection.
PASS optionsLength is 2
PASS selectedOptions.length is 2
PASS selectedOptions[0].text is 'one'
PASS selectedOptions[1].text is 'two'
4) Adding a non selected option should not change the selected options collection.
PASS optionsLength is 3
PASS selectedOptions.length is 1
PASS selectedOptions[0].text is 'one'
5) Adding a selected option should change the selected options collection.
PASS optionsLength is 4
PASS selectedOptions.length is 2
PASS selectedOptions[0].text is 'one'
PASS selectedOptions[1].text is 'five'
6) Unselect an option should update the selected options collection.
PASS optionsLength is 4
PASS selectedOptions.length is 1
PASS selectedOptions[0].text is 'five'
7) Remove an option unselected should not update the selected options collection.
PASS optionsLength is 3
PASS selectedOptions.length is 1
PASS selectedOptions[0].text is 'five'
8) Remove an option selected should update the selected options collection.
PASS optionsLength is 2
PASS selectedOptions.length is 0
9) Change multiple attribute to false should update selectedOptions.
PASS optionsLength is 2
PASS selectedOptions.length is 1
PASS selectedOptions[0].text is 'two'
10) Even with an option disabled selectedOptions should be updated.
PASS optionsLength is 2
PASS selectedOptions.length is 1
PASS selectedOptions[0].text is 'one'
11) Even with select element disabled, the selectedOptions should be updated.
PASS optionsLength is 2
PASS selectedOptions.length is 2
PASS selectedOptions[0].text is 'one'
PASS selectedOptions[1].text is 'two'


PASS successfullyParsed is true

TEST COMPLETE
